Provide tagged dictionary
We can create a very lightweight wrapper around `Dict a b`.

The trick here is that if the thing we're tagging is already comparable,
we can just use the underlying comparable value as the key.

We don't need to reinvent the algorithm;
nor do we need to impose an additional function constraint.
That stuff is taken care of by `Dict` and `Tagged`, respectively.